refactor: rename os-processing module
